Issue #4095: Split and Organize Checkstyle inputs by Test for AbbreviationAsWordInName
