--- src/test/resources/com/puppycrawl/tools/checkstyle/checks/naming/InputAbbreviationAsWordReceiver.java
@@ -1,9 +1,0 @@
-package com.puppycrawl.tools.checkstyle.checks.naming;
-
-public class InputAbbreviationAsWordReceiver {
-    public void foo4(InputAbbreviationAsWordReceiver this) {}
-
-    private class Inner {
-        public Inner(InputAbbreviationAsWordReceiver InputAbbreviationAsWordReceiver.this) {}
-    }
-}
