+++ src/test/resources/com/puppycrawl/tools/checkstyle/checks/naming/abbrevationaswordinname/InputAbbreviationAsWordInNameReceiver.java
@@ -1,0 +1,9 @@
+package com.puppycrawl.tools.checkstyle.checks.naming.abbrevationaswordinname;
+
+public class InputAbbreviationAsWordInNameReceiver {
+    public void foo4(InputAbbreviationAsWordInNameReceiver this) {}
+
+    private class Inner {
+        public Inner(InputAbbreviationAsWordInNameReceiver InputAbbreviationAsWordInNameReceiver.this) {}
+    }
+}
